Fatras: move Bethe-Bloch and Bethe-Heitler
