+++ Fatras/include/ActsFatras/Physics/StandardInteractions.hpp
@@ -10,8 +10,8 @@
 
 #include "ActsFatras/Kernel/ContinuousProcess.hpp"
 #include "ActsFatras/Kernel/InteractionList.hpp"
-#include "ActsFatras/Physics/EnergyLoss/BetheBloch.hpp"
-#include "ActsFatras/Physics/EnergyLoss/BetheHeitler.hpp"
+#include "ActsFatras/Physics/ElectroMagnetic/BetheBloch.hpp"
+#include "ActsFatras/Physics/ElectroMagnetic/BetheHeitler.hpp"
 #include "ActsFatras/Physics/Scattering/Highland.hpp"
 #include "ActsFatras/Selectors/ChargeSelectors.hpp"
 #include "ActsFatras/Selectors/KinematicCasts.hpp"
