+++ Tests/UnitTests/Fatras/Physics/CMakeLists.txt
@@ -1,5 +1,6 @@
 set(unittest_extra_libraries ActsFatras)
 
-add_unittest(FatrasEnergyLoss EnergyLossTests.cpp)
+add_unittest(FatrasBetheBloch BetheBlochTests.cpp)
+add_unittest(FatrasBetheHeitler BetheHeitlerTests.cpp)
 add_unittest(FatrasScattering ScatteringTests.cpp)
 add_unittest(FatrasPhotonConversion PhotonConversionTests.cpp)
